moved the streaming receiver and EQ to dsp.py
